refactor: update DictionaryExtension.cs
