+++ src/WeihanLi.Common/Extensions/ParameterDictionaryExtension.cs
@@ -1,10 +1,35 @@
-﻿using WeihanLi.Common;
+﻿using System.Data;
+using WeihanLi.Common;
 
 // ReSharper disable once CheckNamespace
 namespace WeihanLi.Extensions;
 
 public static class ParameterDictionaryExtension
 {
+    /// <summary>
+    /// IDictionary to dataTable
+    /// </summary>
+    /// <param name="dictionary">IDictionary</param>
+    /// <returns></returns>
+    [RequiresUnreferencedCode("Dictionary value runtime types may not preserve members required by DataColumn.")]
+    public static DataTable ToDataTable(this IDictionary<string, object> dictionary)
+    {
+        Guard.NotNull(dictionary);
+        var dataTable = new DataTable();
+        if (dictionary.Keys.Count == 0)
+        {
+            return dataTable;
+        }
+        dataTable.Columns.AddRange(dictionary.Keys.Select(key => new DataColumn(key, dictionary[key].GetType())).ToArray());
+        foreach (var key in dictionary.Keys)
+        {
+            var row = dataTable.NewRow();
+            row[key] = dictionary[key];
+            dataTable.Rows.Add(row);
+        }
+        return dataTable;
+    }
+    
     /// <summary>
     /// Get param dictionary
     /// </summary>
